fix: include pubky key format in shared targets
