+++ Bitkit/Managers/ContactsManager.swift
@@ -11,46 +11,6 @@ private func stripPubkyPrefix(_ key: String) -> String {
     key.hasPrefix(pubkyPrefix) ? String(key.dropFirst(pubkyPrefix.count)) : key
 }
 
-enum PubkyPublicKeyFormat {
-    private static let rawKeyLength = 52
-    private static let allowedCharacters = Set("ybndrfg8ejkmcpqxot1uwisza345h769")
-
-    static let maximumInputLength = pubkyPrefix.count + rawKeyLength
-
-    static func bounded(_ input: String) -> String {
-        String(input.trimmingCharacters(in: .whitespacesAndNewlines).lowercased().prefix(maximumInputLength))
-    }
-
-    static func normalized(_ input: String) -> String? {
-        let boundedInput = bounded(input)
-        let rawKey = stripPubkyPrefix(boundedInput)
-
-        guard rawKey.count == rawKeyLength else {
-            return nil
-        }
-
-        guard rawKey.allSatisfy({ allowedCharacters.contains($0) }) else {
-            return nil
-        }
-
-        return ensurePubkyPrefix(rawKey)
-    }
-
-    static func matches(_ lhs: String?, _ rhs: String?) -> Bool {
-        guard let lhs = lhs.flatMap(normalized),
-              let rhs = rhs.flatMap(normalized)
-        else {
-            return false
-        }
-
-        return lhs == rhs
-    }
-
-    static func redacted(_ input: String) -> String {
-        (normalized(input) ?? input.trimmingCharacters(in: .whitespacesAndNewlines)).ellipsis(maxLength: 12, style: .end)
-    }
-}
-
 enum AddContactValidationResult: Equatable {
     case empty
     case existingContact
